+++ Utils/FftUtils/sources/rxtd/fft_utils/FFT.cpp
@@ -13,27 +13,27 @@ using rxtd::fft_utils::FFT;
 
 void FFT::setParams(index newSize, std::vector<float> _window) {
 	fftSize = newSize;
-	scalar = 1.0f / float(fftSize);
+	scalar = 1.0f / static_cast<float>(fftSize);
 	window = std::move(_window);
-	kiss.assign(fftSize / 2, false);
+	kiss.assign(static_cast<std::size_t>(fftSize / 2), false);
 
-	inputBuffer.resize(fftSize);
-	outputBuffer.resize(fftSize / 2);
+	inputBuffer.resize(static_cast<size_t>(fftSize));
+	outputBuffer.resize(static_cast<size_t>(fftSize / 2));
 }
 
-double FFT::getDC() const {
+float FFT::getDC() const {
 	return outputBuffer[0].real() * scalar;
 }
 
 float FFT::getBinMagnitude(index binIndex) const {
-	const auto v = outputBuffer[binIndex];
+	const auto v = outputBuffer[static_cast<size_t>(binIndex)];
 	const float square = v.real() * v.real() + v.imag() * v.imag();
 	return std::sqrt(square) * scalar;
 }
 
 void FFT::process(array_view<float> wave) {
 	for (index i = 0; i < fftSize; ++i) {
-		inputBuffer[i] = wave[i] * window[i];
+		inputBuffer[static_cast<size_t>(i)] = wave[i] * window[static_cast<size_t>(i)];
 	}
 
 	kiss.transform_real(inputBuffer.data(), outputBuffer.data());
